move Avatar to User\Avatar, plus bugfixes
