+++ class/Defaults/MergeTag/User/Avatar.php
@@ -4,19 +4,20 @@
  *
  * Requirements:
  * - Trigger property `user_object` or any other passed as
- * `property_name` parameter. Must be an object, preferabely WP_User
+ * `property_name` parameter. Must be an object with a `user_email`
+ * property, preferabely WP_User.
  *
  * @package notification
  */
 
-namespace BracketSpace\Notification\Defaults\MergeTag;
+namespace BracketSpace\Notification\Defaults\MergeTag\User;
 
 use BracketSpace\Notification\Defaults\MergeTag\HtmlTag;
 
 /**
  * Avatar merge tag class
  */
-class AvatarTag extends HtmlTag {
+class Avatar extends HtmlTag {
 
 	/**
 	 * Trigger property to get the user data from
@@ -40,20 +41,17 @@ class AvatarTag extends HtmlTag {
 		$args = wp_parse_args(
 			$params,
 			[
-				'slug'        => 'avatar',
+				'slug'        => 'user_avatar',
 				'name'        => __( 'User avatar', 'notification' ),
 				'description' => get_avatar( get_option( 'admin_email' ) ),
 				'example'     => true,
 				'resolver'    => function( $trigger ) {
 
-                    // check for 'avatar' property, fallback to 'user_email'
-					if ( isset( $trigger->{ $this->property_name }->avatar ) ) {
-                        return get_avatar( $trigger->{ $this->property_name }->avatar );
-                    } else if ( isset( $trigger->{ $this->property_name }->user_email ) ) {
-                        return get_avatar( $trigger->{ $this->property_name }->user_email );
-                    }
+					if ( isset( $trigger->{ $this->property_name }->user_email ) ) {
+						return get_avatar( $trigger->{ $this->property_name }->user_email );
+					}
 
-                    return '';
+					return '';
 				},
 				'group'       => __( 'User', 'notification' ),
 			]
@@ -69,8 +67,7 @@ class AvatarTag extends HtmlTag {
 	 * @return boolean
 	 */
 	public function check_requirements() {
-        return isset( $this->trigger->{ $this->property_name }->avatar )
-            || isset( $this->trigger->{ $this->property_name }->user_email );
+        return isset( $this->trigger->{ $this->property_name }->user_email );
 	}
 
 }
