+++ class/Defaults/Trigger/Comment/CommentTrigger.php
@@ -120,9 +120,9 @@ abstract class CommentTrigger extends Abstracts\Trigger {
 			'group' => sprintf( __( '%s author', 'notification' ), ucfirst( $this->comment_type ) ),
 		] ) );
 
-		$this->add_merge_tag( new MergeTag\AvatarTag( [
-			'slug'  => 'comment_author_avatar',
-			'name'  => __( 'Comment author avatar', 'notification' ),
+		$this->add_merge_tag( new MergeTag\User\Avatar( [
+			'slug'  => 'comment_author_user_avatar',
+			'name'  => __( 'Comment author user avatar', 'notification' ),
 			// translators: comment type author.
 			'group' => sprintf( __( '%s author', 'notification' ), ucfirst( $this->comment_type ) ),
 		] ) );
