+++ class/Defaults/Trigger/Post/PostPublished.php
@@ -131,7 +131,7 @@ class PostPublished extends PostTrigger {
 			'group'         => __( 'Publishing user', 'notification' ),
 		] ) );
 
-		$this->add_merge_tag( new MergeTag\AvatarTag( [
+		$this->add_merge_tag( new MergeTag\User\Avatar( [
 			'slug'          => $this->post_type . '_publishing_user_avatar',
 			// translators: singular post name.
 			'name'          => sprintf( __( '%s publishing user avatar', 'notification' ), $post_name ),
