+++ class/Defaults/Trigger/Post/PostTrashed.php
@@ -126,7 +126,7 @@ class PostTrashed extends PostTrigger {
 			'group'         => __( 'Trashing user', 'notification' ),
 		] ) );
 
-		$this->add_merge_tag( new MergeTag\AvatarTag( [
+		$this->add_merge_tag( new MergeTag\User\Avatar( [
 			'slug'          => $this->post_type . '_trashing_user_avatar',
 			// translators: singular post name.
 			'name'          => sprintf( __( '%s trashing user avatar', 'notification' ), $post_name ),
