+++ class/Defaults/Trigger/User/UserTrigger.php
@@ -37,7 +37,7 @@ abstract class UserTrigger extends Abstracts\Trigger {
 		$this->add_merge_tag( new MergeTag\User\UserLogin() );
 		$this->add_merge_tag( new MergeTag\User\UserEmail() );
 		$this->add_merge_tag( new MergeTag\User\UserRole() );
-		$this->add_merge_tag( new MergeTag\AvatarTag() );
+		$this->add_merge_tag( new MergeTag\User\Avatar() );
 
 		$this->add_merge_tag( new MergeTag\DateTime\DateTime( [
 			'slug' => 'user_registered_datetime',
