+++ readme.txt
@@ -279,7 +279,7 @@ Yes, just activate the debug log in the DEBUGGING section of the plugin settings
 == Changelog ==
 
 = [Next] =
-* [Added] Avatar merge tags.
+* [Added] User Avatar merge tags.
 
 = 6.2.0 =
 * [Fixed] Checkbox in plugin settings now can specify true-ish default value.
